# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Clinic Management System API
  description: |
    Comprehensive API documentation for the Clinic Management System.
    This API provides endpoints for managing patients, appointments, consultations, prescriptions, invoices, and more.
  version: 2.0.0
  contact:
    name: API Support
    email: api-support@sethmedicalclinic.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://api.sethmedicalclinic.com/api/v1
    description: Production server
...
paths:
...
  /patients:
    get:
      tags:
        - Patients
      summary: List patients
      description: Get a list of patients with optional filtering
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of patients
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Patient'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      tags:
        - Patients
      summary: Create patient
      description: Create a new patient record
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientInput'
      responses:
        '201':
          description: Patient created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Patient'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Patient:
      type: object
      properties:
        id:
          type: string
          format: uuid
        patient_number:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        date_of_birth:
          type: string
          format: date
        gender:
          type: string
          enum: [Male, Female, Other]
        phone:
          type: string
        email:
          type: string
        address:
          type: string
    PatientInput:
      type: object
      required:
        - first_name
        - last_name
        - date_of_birth
        - gender
        - phone
      properties:
        first_name:
          type: string
        last_name:
          type: string
        date_of_birth:
          type: string
          format: date
        gender:
          type: string
          enum: [Male, Female, Other]
        phone:
          type: string
        email:
          type: string
        address:
          type: string
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        message:
          type: string
  responses:
    UnauthorizedError:
      description: Unauthorized - Invalid or missing authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: UNAUTHORIZED
            message: Invalid or expired token
    BadRequestError:
      description: Bad Request - Invalid input data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: VALIDATION_ERROR
            message: Invalid input data
...